+++ test/sandbox/app/components/multiple_templates_component.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class MultipleTemplatesComponent < ViewComponent::Base
+  def initialize
+    @items = ["Apple", "Banana", "Pear"]
+  end
+end